requirements.txt
numpy==1.25.0
opencv_python==4.11.0.86
pandas==2.3.2
Pillow==11.3.0
scikit_learn==1.7.2
timm==1.0.19
torch==2.2.2+cu121
torchvision==0.17.2+cu121
yacs==0.1.8
